use -w flag for data removal

+++ remove.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
-while getopts ':d:u:h' flag; do
+WIPE_DATA=false
+while getopts ':d:u:hw' flag; do
   case "${flag}" in
     h)
         echo "Remove a new transmission stack for a user"
@@ -8,10 +9,12 @@ while getopts ':d:u:h' flag; do
         echo "-h, --help                show brief help"
         echo "-d USERSDIR               the directory where to store users"
         echo "-u USERNAME               the username to use"
+        echo "-w                        wipe all downloads data"
         exit 0
         ;;
     d) USERS_DIR="${OPTARG}" ;;
     u) USERNAME="${OPTARG}" ;;
+    w) WIPE_DATA=true ;;
     *) echo "Unexpected option ${flag}" ;;
   esac
 done
@@ -25,19 +28,20 @@ if [ -z ${USERNAME+x} ]; then
     exit 2
 fi
 
-USER_LOCAL_DIR=$USERS_DIR/$USERNAME
-
-if [[ $(whoami) != "root" ]]; then
-    echo "the user removal needs root privileges, please use sudo (with -E to preserve env variables)"
-    exit 3
+if $WIPE_DATA; then
+    if [[ $(whoami) != "root" ]]; then
+        echo "the user removal needs root privileges, please use sudo (with -E to preserve env variables)"
+        exit 3
+    fi
 fi
+
+USER_LOCAL_DIR=$USERS_DIR/$USERNAME
 
 FTP_ID=$(docker ps -f name=^/transmission_stack_ftpd_server$ -q)
 if [ ! -z ${FTP_ID} ]; then
     docker exec -it transmission_stack_ftpd_server pure-pw list | grep $USERNAME && \
         echo "removing ftp config for user $USERNAME" && docker exec -it transmission_stack_ftpd_server /usr/local/bin/remove_user.sh -u $USERNAME
     FTP_USERS=$(docker exec -it transmission_stack_ftpd_server pure-pw list)
-    echo $FTP_USERS
     if [ -z $FTP_USERS ]; then
         echo "removing ftp server as there are no more users"
         docker rm -f transmission_stack_ftpd_server
@@ -49,9 +53,13 @@ if [ ! -z ${SUBFINDER_ID} ]; then
     echo "removing subfinder config for user $USERNAME"
     docker rm -f $SUBFINDER_ID
 fi
+
+if $WIPE_DATA; then
+    userdel $USERNAME
+    rm -rf $USER_LOCAL_DIR
+fi
+
 echo "removing transmission config for user $USERNAME"
-userdel $USERNAME
-rm -rf $USER_LOCAL_DIR
 docker rm -f transmission_stack_transmission_$USERNAME > /dev/null
 
 echo "user $USERNAME successfully removed"
